refactored all but BusinessPeriod in individual files
